+++ templates/part.content.php
@@ -30,10 +30,12 @@
                         </time>
                     </li>
                     <li ng-click="Content.toggleStar(item.id)" class="util" news-stop-propagation>
-                        <button class="star svg" ng-class="{'starred': item.starred}" title="<?php p($l->t('Star')); ?>"></button>
+                        <button class="star svg" ng-hide="item.starred" title="<?php p($l->t('Star article')); ?>"></button>
+                        <button class="starred svg" ng-show="item.starred" title="<?php p($l->t('Unstar article')); ?>"></button>
                     </li>
                     <li ng-click="Content.toggleKeepUnread(item.id)" class="util" news-stop-propagation>
-                        <button class="icon-toggle toggle-keep-unread" ng-class="{'keep-unread': item.keepUnread}" title="<?php p($l->t('Keep article unread')); ?>"></button>
+                        <button class="icon-toggle toggle-keep-unread" ng-hide="item.keepUnread" title="<?php p($l->t('Keep article unread')); ?>"></button>
+                        <button class="icon-toggle toggle-keep-unread keep-unread" ng-show="item.keepUnread" title="<?php p($l->t('Remove keep article unread')); ?>"></button>
                     </li>
                     <li class="util only-in-compact">
                         <a class="external icon-link"